173 any jofile (#377)
* multi job files

* s

* a

* win service

* fix exception

* fix gen job

* fix load-json

* context file

* remove wizard

* final fix

---------

+++ src/Planar.Common/AppSettings/ClusterSettings.cs
@@ -10,7 +10,7 @@ namespace Planar.Common
 
         public TimeSpan HealthCheckInterval { get; set; }
 
-        public short Port { get; set; }
+        public int Port { get; set; }
 
         public bool Clustering { get; set; }
     }
